Improve JSDocs. Seperate out serialized types. Improve README. Other niggles.
