+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "good-flow",
-  "version": "0.0.13",
+  "version": "0.0.14-beta0",
   "description": "Improve how you do Javascript errors",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
