+++ src/good-flow/advice/serialize/types.ts
@@ -1,0 +1,6 @@
+export type SerializeGFAdviceOptions = {
+  /**
+   * @default false
+   */
+  disableColors?: boolean
+}
