+++ src/good-flow/advice/serialized/types.ts
@@ -1,10 +1,3 @@
-export type SerializeGFAdviceOptions = {
-  /**
-   * @default false
-   */
-  disableColors?: boolean
-}
-
 export type SerializedGFTip = string | { msg: string, url?: string }
 
 export type SerializedGFTips = SerializedGFTip | SerializedGFTip[]
