+++ src/index.ts
@@ -2,7 +2,10 @@
  * This file defines the public API of the package. Everything here will be available from
  * the top-level package name when importing as an npm package.
  *
- * E.g. `import { createPackageName, PackageNameOptions } from 'npm-package-name`
+ * E.g. `import ... from 'good-flow`
+ *
+ * Note: Importing anything from here will result in the co-importation of the `colors` package,
+ * which is not valid in browser environments.
  */
 
 export { createGFError } from './good-flow/error'
